+++ quotebot/src/main/java/com/andmark/quotebot/dto/UserDTO.java
@@ -1,6 +1,7 @@
 package com.andmark.quotebot.dto;
 
 import com.andmark.quotebot.domain.enums.UserRole;
+import jakarta.validation.constraints.Size;
 import lombok.*;
 
 @RequiredArgsConstructor
@@ -12,6 +13,7 @@ public class UserDTO {
     private Long id;
     private Long usertgId;
     private String username;
+    @Size(min = 3, max = 20, message = "Password must be between 3 and 20 characters")
     private String password;
     private UserRole role;
     private String nickname;